Unleash the Ghosts!
* Changes:

Initial commit, bringing the GhostCell and its GhostToken to life.

* Motivation:

The promise of writing safe pointer-based container libraries is just
too good to pass up, even if the ergonomics are somewhat reduced.

And whilst everyone could just re-implement it on their own, it is
better to have a single, reviewed, implementation of such a foundational
crate.