Introduce GhostCursor, a cursor for linked GhostCell.
* Changes:

Introduce GhostCursor, with documentation and some examples.

* Motivation:

Navigating linked GhostCell is necessary to move around a data-structure
which itself is necessary to implement the various pieces of
functionality that one wishes of them.

Doing so safely is tricky, due to lifetime concerns, so it is best
abstracted in a common library.

And hopefully, this allows implementing collections in safe Rust.