Introduce TripodList
* Changes:

Introduce TripodList, with its iterator and cursors.

* Motivation:

For the cost of an extra pointer in each node -- the tripod -- it
becomes possible to anchor a node to the stack while still keeping it
linked into the list.

In turn, this allows avoiding the use of the unproven GhostCursor, and
only relying on a single experimental feature: static_rc::lift.

Also, it allows mutably borrowing the list in CursorMut, which in turn
should allow all the edit operations on CursorMut.