Extend GhostBorrowMut, and move into separate module.
* Changes:

- Extend GhostBorrowMut with unsafe, unchecked, version.
- Remove private trait.
- Move into separate module.
- Clarify/extend documentation.

* Motivation:

Having it in the ghost_cell module is unnecessary, so simply move it
into its own module for clarity, and keeping file size reasonable.

Remove the artificial sealed trait. If others wish to write their own,
they should feel free. Hopefully they'll follow the spirit.

Extent the trait with an unsafe, unchecked, version. While not so useful
in the infallible cases, when a run-time check is necessary, the ability
to circumvent it may be useful, so ensure it is available even if it
requires the caller to use unsafe.