Some `GhostCursor` fixes (#3) (#12)
Fix GhostCursor's move_mut and into_parts methods

* Changes

- Fixed `GhostCursor::move_mut` soundness issue as described in #3.
- Switched to returning a mutable token from `GhostCursor::into_parts`, as per #3 again.

* Motivation

As described in #3 , `move_mut` suffered from the possibility of being able to leak the reference, leading to unsoundness (aliasing). A slight change in signature allows catching the issue at compile-time, and a test is added for it.

Also, there is no reason for `into_parts` not to return a mutable token, so it is tweaked to do so.