Constify GhostCell::new
* Changes:

Constify `GhostCell::new`.

* Motivation:

`GhostCell` should be usable in `const` contexts, fully. There really is no reason not to.

For now, we are limited by the state of the language/compiler. It is desirable that ghost-cell compiles on stable, but there seems to be no need to "hold back" for older compilers, so this commit moves things forward as per 1.61 it becomes possible to constify the creation of `GhostCell`.