Fix GhostBorrowMut for shallow projections.
* Changes:

- Check for overlap of memory ranges, not pointer equality.
- Sort unconditionally, for simplicity.

* Motivation:

There is a soundness issue exposed by @noamtashma in #23 which occurs
when using `as_slice_of_cells` as a shallow projection to obtain both a
`GhostCell` of an array or slice and a `GhostCell` of an element within.

By not selecting the first element, the pointers differ, even though the
memory ranges overlap, and thus checking for mere pointer equality is
insufficient.

* Limitation:

This only works for shallow projections, as in projections yielding a
pointer within the memory range of their original; more complex forms of
projection are still not supported, and may never be compatible, as they
would imply checking overlap of reachability graphs, a daunting task.