Add from_slice_of_cells (#30)
* Changes:

- Add from_slice_of_cells to GhostCell<'_, [T]>.

* Motivation:

Symmetric version of as_slice_of_cells to allow back and forth conversions.

* Limitation:

The preferred way would have been an inherent method, but that is not
possible for slices of local types. This leaves us with a trait
implementation or hacky usage of `ops::Index`, neither of which are
particular clear and discoverable.